@@ -1,57 +1,51 @@
-source "http://ruby.taobao.org"
-ruby '1.9.3'
+ruby "2.0.0"
+source "https://rubygems.org"
 
 group :production do
-  gem "unicorn", "~> 4.6.2"
-  gem "pg", "~> 0.15.1"
+  gem "pg", "~> 0.17.1"
+  gem "unicorn", "~> 4.7"
 end
 
 group :development do
-  gem "sqlite3", "~> 1.3.7"
-  gem "coveralls", "~> 0.6.7", require: false
-  gem "pry-debugger", "~> 0.2.2"
-  gem "rspec", "~> 2.13.0"
-  gem "rspec-html-matchers", "~> 0.4.1"
-  gem "rack-test", "~> 0.6.2"
-  gem "shotgun", "~> 0.9"
-  gem "faker", "~> 1.1.2"
-  gem "foreman", "~> 0.63.0"
+  gem "sqlite3", "~> 1.3", ">= 1.3.8"
 end
 
-group :test do
-  gem "coveralls", "~> 0.6.7", require: false
-  gem "pry-debugger", "~> 0.2.2"
-  gem "rspec", "~> 2.13.0"
-  gem "rspec-html-matchers", "~> 0.4.1"
-  gem "rack-test", "~> 0.6.2"
-  gem "shotgun", "~> 0.9"
-  gem "racksh", "~> 1.0.0"
-  gem "faker", "~> 1.1.2"
+group :development, :test do
+  gem "coveralls", "~> 0.7", require: false
+  gem "faker", "~> 1.2"
   gem "foreman", "~> 0.63.0"
+  gem "pry-byebug", "~> 1.2"
+  gem "rack-test", "~> 0.6.2"
+  gem "rspec", "~> 2.14", ">= 2.14.1"
+  gem "rspec-html-matchers", "~> 0.4.3"
+  gem "shotgun", "~> 0.9.0"
 end
 
 group :heroku do
-  gem "excon", "~> 0.25.0"
+  gem "excon", "~> 0.31.0"
   gem "formatador", "~> 0.2.4"
   gem "netrc", "~> 0.7.7"
   gem "rendezvous", "~> 0.0.2"
 end
 
-gem "sinatra", "~> 1.4.2"
-gem "activerecord", "~> 3.2.0"
-gem "sinatra-activerecord", "~> 1.2.2"
-gem "sinatra-flash", "~> 0.3.0"
-gem "sinatra-contrib", github: "sinatra/sinatra-contrib"
-gem "sinatra-assetpack", "~> 0.2.2", require: "sinatra/assetpack"
-gem "i18n", "~> 0.6.4"
-gem "rake", "~> 10.1.0"
-gem "delayed_job_active_record", "~> 0.4.4"
-gem "bcrypt-ruby", "~> 3.0.1"
-gem "will_paginate", "~> 3.0.4"
-gem "feedzirra", "~> 0.2.1"
+gem "activerecord", "~> 4.0"
+# need to work around bug in 4.0.1 https://github.com/rails/arel/pull/216
+gem 'arel', git: 'git://github.com/rails/arel.git', branch: '4-0-stable'
+gem "bcrypt-ruby", "~> 3.1.2"
+gem "delayed_job", "~> 4.0"
+gem "delayed_job_active_record", "~> 4.0"
+gem "feedbag", "~> 0.9.2"
+gem "feedzirra", "~> 0.6.0"
+gem "highline", "~> 1.6", ">= 1.6.20", require: false
+gem "i18n", "~> 0.6.9"
 gem "loofah", github: "swanson/loofah"
-gem "nokogiri", "~> 1.6.0"
-gem "feedbag", github: "dwillis/feedbag"
-gem "highline", "~> 1.6.19", require: false
-gem "thread", "~> 0.0.8"
-gem "racksh", "~> 1.0.0"
+gem "nokogiri", "~> 1.6"
+gem "racksh", "~> 1.0"
+gem "rake", "~> 10.1", ">= 10.1.1"
+gem "sinatra", "~> 1.4", ">= 1.4.4"
+gem "sinatra-assetpack", "~> 0.3.1", require: "sinatra/assetpack"
+gem "sinatra-activerecord", "~> 1.2", ">= 1.2.3"
+gem "sinatra-contrib", ">= 1.4.2"
+gem "sinatra-flash", "~> 0.3.0"
+gem "thread", "~> 0.1.3"
+gem "will_paginate", "~> 3.0", ">= 3.0.5"
