@@ -8,13 +8,16 @@
 # from scratch. The latter is a flawed and unsustainable approach (the more migrations
 # you'll amass, the slower it'll run and the greater likelihood for issues).
 #
-# It's strongly recommended to check this file into your version control system.
+# It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20130905204142) do
+ActiveRecord::Schema.define(version: 20130905204142) do
 
-  create_table "delayed_jobs", :force => true do |t|
-    t.integer  "priority",   :default => 0
-    t.integer  "attempts",   :default => 0
+  # These are extensions that must be enabled in order to support this database
+  enable_extension "plpgsql"
+
+  create_table "delayed_jobs", force: true do |t|
+    t.integer  "priority",   default: 0
+    t.integer  "attempts",   default: 0
     t.text     "handler"
     t.text     "last_error"
     t.datetime "run_at"
@@ -22,43 +25,43 @@ ActiveRecord::Schema.define(:version => 20130905204142) do
     t.datetime "failed_at"
     t.string   "locked_by"
     t.string   "queue"
-    t.datetime "created_at",                :null => false
-    t.datetime "updated_at",                :null => false
+    t.datetime "created_at",             null: false
+    t.datetime "updated_at",             null: false
   end
 
-  add_index "delayed_jobs", ["priority", "run_at"], :name => "delayed_jobs_priority"
+  add_index "delayed_jobs", ["priority", "run_at"], name: "delayed_jobs_priority", using: :btree
 
-  create_table "feeds", :force => true do |t|
+  create_table "feeds", force: true do |t|
     t.string   "name"
-    t.text     "url",          :limit => 255
+    t.text     "url"
     t.datetime "last_fetched"
-    t.datetime "created_at",                  :null => false
-    t.datetime "updated_at",                  :null => false
+    t.datetime "created_at",   null: false
+    t.datetime "updated_at",   null: false
     t.integer  "status"
   end
 
-  add_index "feeds", ["url"], :name => "index_feeds_on_url", :unique => true
+  add_index "feeds", ["url"], name: "index_feeds_on_url", unique: true, using: :btree
 
-  create_table "stories", :force => true do |t|
-    t.text     "title",       :limit => 255
-    t.text     "permalink",   :limit => 255
+  create_table "stories", force: true do |t|
+    t.text     "title"
+    t.text     "permalink"
     t.text     "body"
     t.integer  "feed_id"
-    t.datetime "created_at",                                    :null => false
-    t.datetime "updated_at",                                    :null => false
+    t.datetime "created_at",                  null: false
+    t.datetime "updated_at",                  null: false
     t.datetime "published"
     t.boolean  "is_read"
-    t.boolean  "keep_unread",                :default => false
-    t.boolean  "is_starred",                 :default => false
-    t.text     "entry_id",    :limit => 255
+    t.boolean  "keep_unread", default: false
+    t.boolean  "is_starred",  default: false
+    t.text     "entry_id"
   end
 
-  add_index "stories", ["entry_id", "feed_id"], :name => "index_stories_on_entry_id_and_feed_id", :unique => true
+  add_index "stories", ["entry_id", "feed_id"], name: "index_stories_on_entry_id_and_feed_id", unique: true, using: :btree
 
-  create_table "users", :force => true do |t|
+  create_table "users", force: true do |t|
     t.string   "password_digest"
-    t.datetime "created_at",      :null => false
-    t.datetime "updated_at",      :null => false
+    t.datetime "created_at",      null: false
+    t.datetime "updated_at",      null: false
     t.boolean  "setup_complete"
     t.string   "api_key"
   end
